libubox: update to 1fe93d2 (2026-05-23) snapshot
- ref: https://git.openwrt.org/project/libubox/commit/?id=1fe93d2fefb213ec987763e7e94ce5eaa757bfc3

+++ release/src/router/libubox/avl-cmp.c
@@ -17,7 +17,7 @@
 #include "avl-cmp.h"
 #include "blob.h"
 
-static inline int _min(int v1, int v2)
+static inline size_t _min(size_t v1, size_t v2)
 {
 	return v1 < v2 ? v1 : v2;
 }
@@ -31,7 +31,7 @@ avl_strcmp(const void *k1, const void *k2, void *ptr)
 int
 avl_blobcmp(const void *k1, const void *k2, void *ptr)
 {
-	int len = _min(blob_raw_len(k1), blob_raw_len(k2));
+	size_t len = _min(blob_raw_len(k1), blob_raw_len(k2));
 
 	return memcmp(k1, k2, len);
 }
